Address streaming chunk mark review
